Applied refactored utils_session
