+++ GameAssistant/libs/utils.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+# Maybe to refactor: change "utils.py" to "utils_auth.py"
 from django.http import HttpResponse,HttpResponseRedirect,HttpResponseBadRequest,HttpResponseForbidden
 from django.urls import reverse
 from GameAssistant.models.clients import Client
@@ -6,7 +7,21 @@ from GameAssistant.models.subclients import SubClient
 from GameAssistant.models.games import Game
 from django.contrib.sessions.models import Session
 
-#To refactor: change "utils.py" to "utils_auth.py"
+
+def decorator_example(args):
+    def _decorator_example(func):
+        def wrapper(request, *callback_args, **callback_kwargs):
+            try:
+                if not args:
+                    return func(request, *callback_args, **callback_kwargs)
+                else:
+                    return HttpResponseForbidden("Forbidden!")
+            except Exception as e:
+                return HttpResponseBadRequest('Unknown error while running utils.decorator_example! Details: {0}'.format(e))
+        return wrapper
+    return _decorator_example
+
+
 def check_auth(auth_level):
     def _check_auth(func):
         def wrapper(request, *callback_args, **callback_kwargs):
@@ -81,6 +96,7 @@ def check_auth(auth_level):
         return wrapper
     return _check_auth
 
+
 def game_ongoing(yes_or_no, auth_level):
     def _game_ongoing(func):
         @check_auth(auth_level)
@@ -126,18 +142,6 @@ def game_ongoing(yes_or_no, auth_level):
         return wrapper
     return _game_ongoing
 
-def decorator_example(args):
-    def _decorator_example(func):
-        def wrapper(request, *callback_args, **callback_kwargs):
-            try:
-                if not args:
-                    return func(request, *callback_args, **callback_kwargs)
-                else:
-                    return HttpResponseForbidden("Forbidden!")
-            except Exception as e:
-                return HttpResponseBadRequest('Unknown error while running utils.decorator_example! Details: {0}'.format(e))
-        return wrapper
-    return _decorator_example
 
 #Works for both super_user and sub_user
 def user_is_seated(user_id, game):
@@ -145,61 +149,3 @@ def user_is_seated(user_id, game):
         if seat.user_id == user_id:
             return True
     return False
-
-#Used for:
-#Get the client_id when being logged in as a super-user;
-#Get the client_id of its super user when working as a sub-user;
-@check_auth('user')
-def get_client_id_from_session(request):
-    try:
-        sessionid = request.COOKIES.get('sessionid')
-        session = Session.objects.get(session_key=sessionid)
-        client_id = session.get_decoded().get('client_id')
-        if not client_id:
-            subclient_id = session.get_decoded().get('subclient_id')
-            client_id = subclient_id.split('@',1)[-1]
-            if not client_id: 
-                return HttpResponseBadRequest('Unknown error happened! Might be due to expired COOKIES or illegal subclient_id!')
-        return client_id
-
-    except Exception as e:
-        return HttpResponseBadRequest('Unknown error while running utils.get_client_id_from_session! Details: {0}'.format(e))
-
-#Used for:
-#Get the client_id when being logged in as a super-user;
-#Get the subclient_id when working as a sub-user;
-@check_auth('user')
-def get_user_id_from_session(request):
-    try:
-        sessionid = request.COOKIES.get('sessionid')
-        session = Session.objects.get(session_key=sessionid)
-        client_id = session.get_decoded().get('client_id')
-        if not client_id:
-            subclient_id = session.get_decoded().get('subclient_id')
-            return subclient_id
-        return client_id
-
-    except Exception as e:
-        return HttpResponseBadRequest('Unknown error while running utils.get_user_id_from_session! Details: {0}'.format(e))
-
-#Used for:
-#Get the client_name when being logged in as a super-user;
-#Get the subclient_name when working as a sub-user;
-@check_auth('user')
-def get_user_name_from_session(request):
-    try:
-        sessionid = request.COOKIES.get('sessionid')
-        session = Session.objects.get(session_key=sessionid)
-        client_id = session.get_decoded().get('client_id')
-        if not client_id:
-            subclient_id = session.get_decoded().get('subclient_id')
-            client_id = subclient_id.split('@',1)[-1]
-            subclient = Client.objects(client_id=client_id).first().subclients.filter(subclient_id=subclient_id).first()
-            return subclient.subclient_name
-        else:
-            client = Client.objects(client_id=client_id).first()
-            return client.client_name
-
-    except Exception as e:
-        return HttpResponseBadRequest('Unknown error while running utils.get_user_name_from_session! Details: {0}'.format(e))
-
